server/requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient
djangorestframework-simplejwt
python-dotenv